refactor: consolidate language of AuxTimer
