+++ apps/client/src/features/control/playback/PlaybackControl.tsx
@@ -3,7 +3,7 @@ import { Playback } from 'ontime-types';
 import { usePlaybackControl } from '../../../common/hooks/useSocket';
 
 import AddTime from './add-time/AddTime';
-import { ExtraTimer } from './extra-timer/ExtraTimer';
+import { AuxTimer } from './aux-timer/AuxTimer';
 import PlaybackButtons from './playback-buttons/PlaybackButtons';
 import PlaybackTimer from './playback-timer/PlaybackTimer';
 
@@ -22,7 +22,7 @@ export default function PlaybackControl() {
         numEvents={data.numEvents}
         selectedEventIndex={data.selectedEventIndex}
       />
-      <ExtraTimer />
+      <AuxTimer />
     </div>
   );
 }
