+++ packages/types/src/definitions/runtime/AuxTimer.type.ts
