+++ packages/types/src/definitions/runtime/RuntimeStore.type.ts
@@ -1,5 +1,5 @@
 import type { OntimeEvent } from '../core/OntimeEvent.type.js';
-import type { SimpleTimerState } from './ExtraTimer.type.js';
+import type { SimpleTimerState } from './AuxTimer.type.js';
 import type { MessageState } from './MessageControl.type.js';
 import type { Runtime } from './Runtime.type.js';
 import type { TimerState } from './TimerState.type.js';
